代码/adaptor/requirements.txt
ahocorapy==1.6.1
alnlp==1.0.0rc27
antlr4-python3-runtime==4.9.3
apache-libcloud==2.3.0
asgiref==3.4.1
biopython==1.79
boto==2.48.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
colorlog==4.6.2
configparser==5.2.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
dataclasses==0.8
DBUtils==2.0.2
Django==3.2.14
django-cors-headers==3.10.1
django-filter==0.9.2
django-stubs==1.9.0
django-stubs-ext==0.5.0
djangorestframework==3.13.1
djangorestframework-stubs==1.4.0
docker==2.0.0
docker-pycreds==0.4.0
filelock==3.0.12
Flask==2.0.1
future==0.18.2
gevent==21.8.0
greenlet==1.1.1
hanlp==2.1.0a55
hanlp-common==0.0.9
hanlp-downloader==0.0.24
hanlp-trie==0.0.2
httplib2==0.11.3
huggingface-hub==0.0.12
idna==3.2
importlib-metadata==4.7.1
iso8601==1.1.0
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
manageiq-client==0.6.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mypy==0.961
mypy-extensions==0.4.3
mysql-connector-python==8.0.6
numpy==1.19.5
packaging==21.0
pandas==1.1.5
parsedatetime==2.6
phrasetree==0.0.8
Pillow==8.4.0
plotly==5.11.0
protobuf==3.19.4
psutil==5.4.5
pycparser==2.20
pyeda @ file:///D:/Python34/Lib/site-packages/pyeda-0.28.0-cp36-cp36m-win_amd64.whl
PyJWT==1.7.1
PyMySQL==1.0.2
pynvml==11.0.0
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.2
pytz==2021.1
pywin32==305
PyYAML==5.4.1
railroad-diagrams==1.1.1
regex==2021.8.21
reload==0.9
requests==2.26.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
sentencepiece==0.1.96
simplejson==3.18.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
tenacity==8.1.0
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.10.3
toml==0.10.2
tomli==1.2.3
toposort==1.5
torch==1.9.0
tqdm==4.56.0
transformers==4.9.2
typed-ast==1.5.4
types-pytz==2022.1.1
types-PyYAML==6.0.9
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==1.26.6
wait-for==1.2.0
watchdog==2.1.6
websocket-client==1.3.1
Werkzeug==2.0.1
wget==3.2
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
